requirements.txt
Django
python-decouple
psycopg2
djangorestframework
django-cors-headers

